reorganize files in the pkg
